9.一键生成QQ个人历史报告/requirement.txt
requests
PyExecJS
Pillow